+++ types/types.go
@@ -16,7 +16,7 @@ type Config struct {
 	BootstrapURL                string
 	Credentials                 UserCredentials
 	CleanupLoopTime             time.Duration
-	ReviveLoopTime              time.Duration
+	CGroupV2                    bool
 	DSMax                       int
 	DSCPU                       float64
 	DSMEM                       float64
@@ -67,6 +67,7 @@ type Config struct {
 	M3SStatus                   M3SStatus
 	MesosSandboxVar             string
 	Principal                   string
+	ReviveLoopTime              time.Duration
 	RedisServer                 string
 	RedisPassword               string
 	RedisDB                     int