+++ changelog.md
@@ -15,6 +15,7 @@
 - Fix scale up performance
 - Add support for cri-docker (enable with K3S_DOCKER=true)
 - Add support for MySQL TLS datastore communication
+- Add support for CGroupV2 (bool env CGROUP_V2, default false)
 
 ## v0.3.0